+++ src/collections/queue/index.ts
@@ -1,9 +1,7 @@
-
-export * from './QueueTypes.js';
-export type { IQueueMutable, IQueueMutableWithEvents } from './IQueueMutable.js';
-export type { IQueueImmutable } from './IQueueImmutable.js';
-export { type IPriorityQueueMutable } from './IPriorityQueueMutable.js';
-export { type QueueOpts } from './QueueTypes.js';
+export type * from './IQueueMutable.js';
+export type * from './IQueueImmutable.js';
+export type * from './IPriorityQueueMutable.js';
+export type * from './QueueTypes.js';
 export { immutable } from './QueueImmutable.js';
 export { mutable, QueueMutable } from './QueueMutable.js';
 export { priority } from './PriorityMutable.js';