+++ src/collections/set/index.ts
@@ -5,3 +5,4 @@ export { immutable } from './SetImmutable.js';
 export { type ISetMutable } from './ISetMutable.js';
 export { type ISetImmutable } from './ISetImmutable.js';
 export * from './MassiveSet.js';
+export type * from './Types.js';