+++ src/collections/stack/index.ts
@@ -1,9 +1,6 @@
 export type { IStack } from './IStack.js';
 export type { IStackMutable } from './IStackMutable.js';
 export type { IStackImmutable } from './IStackImmutable.js';
-
-// ✔ Unit tested!
-
-
 export { immutable } from './StackImmutable.js';
 export { mutable } from './StackMutable.js';
+export type * from './Types.js';